separate bot and bot_advisor, as first step towards splitting PR in 2 (or 3)
